chore(refactor): move schema clean up into separate class
