+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/SchemaGenerator.java
@@ -18,25 +18,19 @@ package com.github.victools.jsonschema.generator;
 
 import com.fasterxml.classmate.ResolvedType;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.github.victools.jsonschema.generator.impl.AttributeCollector;
 import com.github.victools.jsonschema.generator.impl.DefinitionKey;
+import com.github.victools.jsonschema.generator.impl.SchemaCleanUpUtils;
 import com.github.victools.jsonschema.generator.impl.SchemaGenerationContextImpl;
 import com.github.victools.jsonschema.generator.impl.TypeContextFactory;
 import java.lang.reflect.Type;
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 import java.util.TreeMap;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Consumer;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
 /**
  * Generator for JSON Schema definitions via reflection based analysis of a given class.
@@ -89,12 +83,11 @@ public class SchemaGenerator {
         }
         ObjectNode mainSchemaNode = generationContext.getDefinition(mainKey);
         jsonSchemaResult.setAll(mainSchemaNode);
+        SchemaCleanUpUtils cleanUpUtils = new SchemaCleanUpUtils(this.config);
         if (this.config.shouldCleanupUnnecessaryAllOfElements()) {
-            String allOfTagName = this.config.getKeyword(SchemaKeyword.TAG_ALLOF);
-            this.finaliseSchemaParts(jsonSchemaResult, nodeToCheck -> this.mergeAllOfPartsIfPossible(nodeToCheck, allOfTagName));
+            cleanUpUtils.reduceAllOfNodes(jsonSchemaResult);
         }
-        String anyOfTagName = this.config.getKeyword(SchemaKeyword.TAG_ANYOF);
-        this.finaliseSchemaParts(jsonSchemaResult, nodeToCheck -> this.reduceAnyOfWrappersIfPossible(nodeToCheck, anyOfTagName));
+        cleanUpUtils.reduceAnyOfNodes(jsonSchemaResult);
 
         return jsonSchemaResult;
     }
@@ -200,195 +193,5 @@ public class SchemaGenerator {
                 .replaceAll("<", "(")
                 .replaceAll(">", ")");
         return uriCompatibleName;
-    }
-
-    /**
-     * Collect names of schema tags that may contain sub-schemas, i.e. {@link SchemaKeyword#TAG_ADDITIONAL_PROPERTIES} and
-     * {@link SchemaKeyword#TAG_ITEMS}.
-     *
-     * @return names of eligible tags as per the designated JSON Schema version
-     * @see #discardUnnecessaryAllOfWrappers(ObjectNode)
-     */
-    private Set<String> getTagNamesContainingSchema() {
-        SchemaVersion schemaVersion = this.config.getSchemaVersion();
-        return Stream.of(SchemaKeyword.TAG_ADDITIONAL_PROPERTIES, SchemaKeyword.TAG_ITEMS)
-                .map(keyword -> keyword.forVersion(schemaVersion))
-                .collect(Collectors.toSet());
-    }
-
-    /**
-     * Collect names of schema tags that may contain arrays of sub-schemas, i.e. {@link SchemaKeyword#TAG_ALLOF}, {@link SchemaKeyword#TAG_ANYOF} and
-     * {@link SchemaKeyword#TAG_ONEOF}.
-     *
-     * @return names of eligible tags as per the designated JSON Schema version
-     * @see #discardUnnecessaryAllOfWrappers(ObjectNode)
-     */
-    private Set<String> getTagNamesContainingSchemaArray() {
-        SchemaVersion schemaVersion = this.config.getSchemaVersion();
-        return Stream.of(SchemaKeyword.TAG_ALLOF, SchemaKeyword.TAG_ANYOF, SchemaKeyword.TAG_ONEOF)
-                .map(keyword -> keyword.forVersion(schemaVersion))
-                .collect(Collectors.toSet());
-    }
-
-    /**
-     * Collect names of schema tags that may contain objects with sub-schemas as values, i.e. {@link SchemaKeyword#TAG_PATTERN_PROPERTIES} and
-     * {@link SchemaKeyword#TAG_PROPERTIES}.
-     *
-     * @return names of eligible tags as per the designated JSON Schema version
-     * @see #discardUnnecessaryAllOfWrappers(ObjectNode)
-     */
-    private Set<String> getTagNamesContainingSchemaObject() {
-        SchemaVersion schemaVersion = this.config.getSchemaVersion();
-        return Stream.of(SchemaKeyword.TAG_PATTERN_PROPERTIES, SchemaKeyword.TAG_PROPERTIES)
-                .map(keyword -> keyword.forVersion(schemaVersion))
-                .collect(Collectors.toSet());
-    }
-
-    /**
-     * Iterate through a generated and fully populated schema and remove extraneous {@link SchemaKeyword#TAG_ANYOF} nodes, where one entry of the
-     * array is again a {@link SchemaKeyword#TAG_ANYOF} wrapper and nothing else. This makes for more readable schemas being generated but has the
-     * side-effect that any manually added {@link SchemaKeyword#TAG_ANYOF} (e.g. through a custom definition of attribute overrides) may be removed as
-     * well if it isn't strictly speaking necessary.
-     *
-     * @param schemaNode generated schema to clean-up
-     * @param performCleanUpOnSingleSchemaNode clean up task to execute before looking for deeper nested sub-schemas for which to apply the same
-     */
-    private void finaliseSchemaParts(ObjectNode schemaNode, Consumer<ObjectNode> performCleanUpOnSingleSchemaNode) {
-        List<ObjectNode> nextNodesToCheck = new ArrayList<>();
-        Consumer<JsonNode> addNodeToCheck = node -> {
-            if (node instanceof ObjectNode) {
-                nextNodesToCheck.add((ObjectNode) node);
-            }
-        };
-        nextNodesToCheck.add(schemaNode);
-        Optional.ofNullable(schemaNode.get(this.config.getKeyword(SchemaKeyword.TAG_DEFINITIONS)))
-                .filter(definitions -> definitions instanceof ObjectNode)
-                .ifPresent(definitions -> ((ObjectNode) definitions).forEach(addNodeToCheck));
-
-        Set<String> tagsWithSchemas = this.getTagNamesContainingSchema();
-        Set<String> tagsWithSchemaArrays = this.getTagNamesContainingSchemaArray();
-        Set<String> tagsWithSchemaObjects = this.getTagNamesContainingSchemaObject();
-        do {
-            List<ObjectNode> currentNodesToCheck = new ArrayList<>(nextNodesToCheck);
-            nextNodesToCheck.clear();
-            for (ObjectNode nodeToCheck : currentNodesToCheck) {
-                performCleanUpOnSingleSchemaNode.accept(nodeToCheck);
-                tagsWithSchemas.stream().map(nodeToCheck::get).forEach(addNodeToCheck);
-                tagsWithSchemaArrays.stream()
-                        .map(nodeToCheck::get)
-                        .filter(possibleArrayNode -> possibleArrayNode instanceof ArrayNode)
-                        .forEach(arrayNode -> arrayNode.forEach(addNodeToCheck));
-                tagsWithSchemaObjects.stream()
-                        .map(nodeToCheck::get)
-                        .filter(possibleObjectNode -> possibleObjectNode instanceof ObjectNode)
-                        .forEach(objectNode -> objectNode.forEach(addNodeToCheck));
-            }
-        } while (!nextNodesToCheck.isEmpty());
-    }
-
-    /**
-     * Check whether the given schema node and its {@link SchemaKeyword#TAG_ALLOF} elements (if there are any) are distinct. If yes, remove the
-     * {@link SchemaKeyword#TAG_ALLOF} node and merge all its elements with the given schema node instead.
-     * <br>
-     * This makes for more readable schemas being generated but has the side-effect that manually added {@link SchemaKeyword#TAG_ALLOF} (e.g. from a
-     * custom definition or attribute overrides) may be removed as well if it isn't strictly speaking necessary.
-     *
-     * @param schemaNode single node representing a sub-schema to consolidate contained {@link SchemaKeyword#TAG_ALLOF} for (if present)
-     * @param allOfTagName name of the {@link SchemaKeyword#TAG_ALLOF} in the designated JSON Schema version
-     */
-    private void mergeAllOfPartsIfPossible(JsonNode schemaNode, String allOfTagName) {
-        if (!(schemaNode instanceof ObjectNode)) {
-            return;
-        }
-        JsonNode allOfTag = schemaNode.get(allOfTagName);
-        if (!(allOfTag instanceof ArrayNode)) {
-            return;
-        }
-        allOfTag.forEach(part -> this.mergeAllOfPartsIfPossible(part, allOfTagName));
-
-        List<JsonNode> allOfElements = new ArrayList<>();
-        allOfTag.forEach(allOfElements::add);
-        if (allOfElements.stream().anyMatch(part -> !(part instanceof ObjectNode) && !part.asBoolean())) {
-            return;
-        }
-        List<ObjectNode> parts = allOfElements.stream()
-                .filter(part -> part instanceof ObjectNode)
-                .map(part -> (ObjectNode) part)
-                .collect(Collectors.toList());
-
-        // collect all defined attributes from the separate parts and check whether there are incompatible differences
-        Map<String, List<JsonNode>> fieldsFromAllParts = Stream.concat(Stream.of(schemaNode), parts.stream())
-                .flatMap(part -> StreamSupport.stream(((Iterable<Map.Entry<String, JsonNode>>) () -> part.fields()).spliterator(), false))
-                .collect(Collectors.groupingBy(Map.Entry::getKey, Collectors.mapping(Map.Entry::getValue, Collectors.toList())));
-        if ((this.config.getSchemaVersion() == SchemaVersion.DRAFT_6 || this.config.getSchemaVersion() == SchemaVersion.DRAFT_7)
-                && fieldsFromAllParts.containsKey(this.config.getKeyword(SchemaKeyword.TAG_REF))) {
-            // in Draft 7, any other attributes besides the $ref keyword were ignored
-            return;
-        }
-        String ifTagName = this.config.getKeyword(SchemaKeyword.TAG_IF);
-        for (Map.Entry<String, List<JsonNode>> fieldEntries : fieldsFromAllParts.entrySet()) {
-            if (fieldEntries.getValue().size() == 1) {
-                // no conflicts, no further checks
-                continue;
-            }
-            if (ifTagName.equals(fieldEntries.getKey())) {
-                // "if"/"then"/"else" tags should remain isolated in their sub-schemas
-                return;
-            }
-            int offset;
-            if (!allOfTagName.equals(fieldEntries.getKey())) {
-                offset = 0;
-            } else if (fieldEntries.getValue().size() == 2) {
-                // we can ignore the "allOf" tag in the target node (the one we are trying to remove here)
-                continue;
-            } else {
-                offset = 1;
-            }
-            if (!fieldEntries.getValue().stream().skip(offset + 1).allMatch(fieldEntries.getValue().get(offset)::equals)) {
-                // different values for the same tag: be conservative for now and not merge anything
-                // later, we may want to decide based on the tag how to merge them (e.g. take the highest "minLength" and the lowest "maximum")
-                return;
-            }
-        }
-        // all attributes are either distinct or have equal values in all occurrences
-        ObjectNode schemaObjectNode = (ObjectNode) schemaNode;
-        schemaObjectNode.remove(allOfTagName);
-        parts.forEach(schemaObjectNode::setAll);
-    }
-
-    /**
-     * Check whether the given schema node contains a {@link SchemaKeyword#TAG_ANYOF} element which in turn contains an entry with only another
-     * {@link SchemaKeyword#TAG_ANYOF} inside. If yes, move the entries from the inner array up to the outer one.
-     * <br>
-     * This makes for more readable schemas being generated but has the side-effect that manually added {@link SchemaKeyword#TAG_ANYOF} entries (e.g.
-     * from a custom definition or attribute overrides) may be removed as well if it isn't strictly speaking necessary.
-     *
-     * @param schemaNode single node representing a sub-schema to consolidate contained {@link SchemaKeyword#TAG_ANYOF} for (if present)
-     * @param anyOfTagName name of the {@link SchemaKeyword#TAG_ANYOF} in the designated JSON Schema version
-     */
-    private void reduceAnyOfWrappersIfPossible(JsonNode schemaNode, String anyOfTagName) {
-        if (!(schemaNode instanceof ObjectNode)) {
-            return;
-        }
-        JsonNode anyOfTag = schemaNode.get(anyOfTagName);
-        if (!(anyOfTag instanceof ArrayNode)) {
-            return;
-        }
-        anyOfTag.forEach(part -> this.reduceAnyOfWrappersIfPossible(part, anyOfTagName));
-
-        for (int index = anyOfTag.size() - 1; index > -1; index--) {
-            JsonNode arrayEntry = anyOfTag.get(index);
-            if (!(arrayEntry instanceof ObjectNode) || arrayEntry.size() != 1) {
-                continue;
-            }
-            JsonNode nestedAnyOf = arrayEntry.get(anyOfTagName);
-            if (!(nestedAnyOf instanceof ArrayNode)) {
-                continue;
-            }
-            ((ArrayNode) anyOfTag).remove(index);
-            for (int nestedEntryIndex = nestedAnyOf.size() - 1; nestedEntryIndex > -1; nestedEntryIndex--) {
-                ((ArrayNode) anyOfTag).insert(index, nestedAnyOf.get(nestedEntryIndex));
-            }
-        }
     }
 }
